# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simply backend app",
        "title": "Split bill backend",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:4949",
    "basePath": "/api/v1",
    "paths": {
        "/bills": {
            "get": {
                "description": "Возвращает список счетов для указанного пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bills"
                ],
                "summary": "Получение всех счетов по ID пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список счетов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Bills"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные счета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bills"
                ],
                "summary": "Редактирование счета",
                "parameters": [
                    {
                        "description": "Обновленные данные счета",
                        "name": "bill",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Bills"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Счет успешно обновлен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новый счет для пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bills"
                ],
                "summary": "Создание нового счета",
                "parameters": [
                    {
                        "description": "Данные счета",
                        "name": "bill",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Bills"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Счет успешно создан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет счет по указанному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bills"
                ],
                "summary": "Удаление счета по ID",
                "parameters": [
                    {
                        "description": "ID счета для удаления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.BillDel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Счет успешно удален",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.BillDel": {
            "type": "object",
            "properties": {
                "bill_id": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "entity.Bills": {
            "type": "object",
            "properties": {
                "bill_id": {
                    "type": "string",
                    "example": "1"
                },
                "bill_title": {
                    "type": "string",
                    "example": "Weekend with friends"
                },
                "user_id": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}